Restriction of the commands during RPC action
Until now, when using RPC action, there were no restrictions on the
commands you could use. However, with RPC actions, only certain commands
should be used, and using others could lead to unexpected problems. This
problem was first noticed when attempting to exit the RPC action
context, as described in the bug link in the bottom. The issue has been
fixed, and this change provides restrictions to additional commands.

This update introduces a new function called `enforceRpcRestrictions`,
which checks if a command is allowed during RPC action.

The following commands are now blocked entirely during RPC action:
- commit
- discard
- copy
- switch
- prepare

The following commands are blocked when leaving or making changes
outside the RPC action:
- create
- set
- delete
- cd
- move

Other commands were not changed.

Bug: https://tree.taiga.io/project/jktjkt-netconf-cli/issue/223
Bug: https://github.com/CESNET/netconf-cli/issues/19
Change-Id: I5bf935bb8eabeea313b506617d1f43274eb77cdf
Signed-off-by: Jan Kundrát <jan.kundrat@cesnet.cz>