Add a real path computation
Many commands need a path as an argument. This path can be absolute or
relative. The key is to understand where the command will operate. For example,
when using `cd`, we need to know where we'll end up after running the command.
And with `set`, we want to know what will be modified. This change creates a
general function called realPath() that takes the current working directory
(cwd) and the given path as arguments, and then gives back the final path.
Doing this makes the code easier to read and more compact because we'll use it
several times in the next changes.

Change-Id: Id940062525b64eed1de438407ff8ae2cfbcdde6d