yang-cli: automatically extend the on-disk search path
When asking for a file to load as a YANG module, automatically add its
enclosing directory to the list of search paths. That way, if a module
imports other modules from the same directory, no `-s` parameter will be
needed.

There should be no harm in adding the same path over and over again, so
there's no point in tracking what paths have been already enabled.

Change-Id: I336efb7b5e6f923e7d9f82741ee93a88bb32bdbe