YANG schema: fix addSchemaDirectory()
A method that's called "addFoo" should add something to foo. However,
this one was *overriding* the directory list because that's the way how
libyang the C library works.

Fix it by making the interface a bit more type safe (hi
std::filesystem::path) as well as by maintaining a "cache" of paths that
were added so far. Note that because it's a std::set, this thing does
not maintain the insertion order, and therefore if someone has the same
YANG model in different directories with different content, unexpected
things will happen.

Also change the interface of the yang-cli program so that the `-s`
option can be now specified several times. The old, undocumented version
with multiple paths separated by colon should still work because of the
way how the entire argument is serialized for libyang, but that's just a
side effect and nobody cares about that.

Change-Id: Iad40696b0e133c60e0c0a1512da851650f6ac8a0