yang-cli: allow parsing modules with "mandatory" features
Some YANG modules, such as the CzechLight ROADM ones, require at least
"some" YANG-level features to be set. For example, there's a pair of
features A and B which affect the set allowed values for an enum. When
none of these features is enabled, the module would fail to load because
of no allowed values for a (used) enum.

Fix that by building the list of features-to-enable up front, and then
requesting these features to be enabled at module load time.

Change-Id: I927eee0b05b7c27972af5a3431f188cf51d09ff9