build: use a less confusing FindPkgConfig syntax
The docs [1] say that the first parameter is the CMake variable to use,
then the keywords like `IMPORTED_TARGET` follow, and at the very end we
have a list of pkg-config bits to search for. Let's not interleave the
list of .pc files with CMake-level keywords; that's rather confusing (I
know I used to read this as "ok, this `IMPORTED_TARGET foo` says that
it's going to create the `PkgConfig::foo`", but that's not the case at
all).

[1] https://cmake.org/cmake/help/latest/module/FindPkgConfig.html

Change-Id: Ic01984ed79a9ea2e1491ceb089c30f5dedbeb755