standalone YANG: disable loading the ietf-yang-library
When working in the in-memory, YANG-only standalone mode, do not load
the ietf-yang-library YANG module. If it is loaded, there are global
mandatory nodes such as `content-id` which prevent successful
validation.

This will be a problem once we merge the dependency stack with an update
for https://gerrit.cesnet.cz/c/CzechLight/libyang-cpp/+/7887 because
that one started throwing exceptions upon validation failure.

As a quick fix, simply don't load that module by default. I think that
people can still load it explicitly, which might be useful for some
quick prototyping. That said, this part of the app needs more work,
there's no support for selectively enabling features, etc.

Change-Id: I776a10eb6059727eb022dd30768a5962ffc61ad0
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/7990