CI: remove explicit timeout for the ARM build
The idea was to ensure that we have "enough time" to finalize the build
and test by giving more time to this project compared to the other C++
ones. However, this is probably not a good idea for the ARM one given
that other projects typically finish in roughly 30min already. So, let's
use the global default for that one instead.

Fixes: 7d7cf4b4 CI: allow longer builds
Change-Id: I9f677c22d7912c2fb202c581f5fe12c36460ce0d
Reported-by: Tomáš Pecka <Tomas.Pecka@cesnet.cz>